Performance fix - Control provider refactor
