+++ src/client/device/device.tsx
@@ -14,6 +14,7 @@ import { RESX } from '../strings';
 
 import { DeviceContext } from '../context/deviceContext';
 import { AppContext } from '../context/appContext';
+import { ControlProvider } from '../context/controlContext';
 
 export function Device() {
 
@@ -22,7 +23,7 @@ export function Device() {
 
     return <>{Object.keys(deviceContext.device).length > 0 ?
         <div className='device'>
-            <div className='device-toolbar'><DeviceToolbar /></div>
+            <div className='device-toolbar'><ControlProvider><DeviceToolbar /></ControlProvider></div>
             <div className='device-title'><DeviceTitle /></div>
             <div className='device-commands'><DeviceCommands /></div>
