+++ src/client/modals/consoleInspector.scss
